Add bot difficulties for players to select
Added a easy and hard mode for the bots

Easy mode is still a random selection based on
available board spaces.

Hard mode uses the minimax algorithm to find the best
space available on the board.

Added easy and hard buttons that appear after a bot
has been selected.

Difficulty setting is applied to the bot player object